Extract all gallery schema types to a separate file
